Added in anims
I've added in the animations and made a few adjustments to the bg of the level. Please @Tobi make the camera smaller - the player is still far too small for the screen. Also note to Damian: Please make sure all frames and animations are in the same ordering layer, and set the layer to "Player" and set them on layer 2, so that they do not dissapear behind platforms. Also note to Tobi: there's a glitch where sometimes the player will spawn aaalll they way back to the first respawn point which can bequite irritating. Idk if thats just on my side tho XD